Move VistedTable class from HNSW.h to AuxIndexStructures.h
